fix for incorrect use of size_t for a value that will decrement below 0
Imported-from: https://subversion.assembla.com/svn/AvP/branches/workbranch/trunk@139

+++ 3dc/console.cpp
@@ -372,7 +372,7 @@ void Con_Draw()
 	int xOffset = 0;
 
 	// draw all the lines of text
-	for (size_t i = rows; i >= 0; i--, y -= CHAR_HEIGHT)
+	for (int32_t i = rows; i >= 0; i--, y -= CHAR_HEIGHT)
 	{
 		xOffset = 0;
 		charWidth = 0;
